+++ Assets/Script/Manager/Base/BaseProduce.cs
@@ -7,7 +7,7 @@ public abstract class BaseProduce : BaseBuild
 {
     
     // 生产速度
-    protected int productionSpeed;
+    protected float productionSpeed;
     // 生产计时
     private float m_productionTimer;
     protected override void Awake()
